requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
openai==1.30.1
python-dotenv==1.0.1
